requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
dj-database-url==2.3.0
Django==4.2
django-ckeditor==6.7.2
django-crispy-forms==2.3
django-js-asset==3.0.1
django-tinymce==4.1.0
environs==14.1.1
gunicorn==23.0.0
marshmallow==3.26.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
whitenoise==6.9.0
